+++ test/endpoint.js
@@ -36,6 +36,12 @@ describe('inflectors', function() {
   });
 });
 
+describe('keyers', function() {
+
+  require('./keyers/omission.js');
+  require('./keyers/skeleton.js');
+});
+
 describe('classification', function() {
   require('./classification/naive-bayes.js');
   require('./classification/perceptron.js');
@@ -106,11 +112,9 @@ describe('phonetics', function() {
   require('./phonetics/metaphone.js');
   require('./phonetics/mra.js');
   require('./phonetics/nysiis.js');
-  require('./phonetics/omission.js');
   require('./phonetics/onca.js');
   require('./phonetics/phonex.js');
   require('./phonetics/roger-root.js');
-  require('./phonetics/skeleton.js');
   require('./phonetics/soundex.js');
   require('./phonetics/statcan.js');