+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Providing index of item in some `clustering/record-linkage` callbacks.
 * Adding `merge` option to `clustering/record-linkage/key-collision`.
 * Adding the `keyers/fingerprint` namespace back.
+* Moving `phonetics/omission` & `phonetics/skeleton` back to the `keyers` namespace.
 
 ## 0.15.0